Refactors modules structure
